Fix wrong types (#16)
* Remove deprecated function calls

time.sleep_ms is deprecated replace with wait

* Change conversion of time

Use functions instead of number for time conversion

* Change time conversion

Add multiplier for duartions

* Fix misspell

* Fix wrong type

* Fix wrong type

* well, it is what it is

+++ rest/rest.v
@@ -34,8 +34,8 @@ pub fn (mut rest REST) do(req http.Request) ?http.Response {
 		mut obj := json.raw_decode(resp.text) ?
 		mut tmr := TooManyRequests{}
 		tmr.from_json(obj)
-		rest.rl.global = time.utc().unix_time_milli() + u64(tmr.retry_after * 1000)
-		time.wait(tmr.retry_after * 1000 * time.millisecond)
+		rest.rl.global = time.utc().unix_time_milli() + u64(tmr.retry_after) * 1000
+		time.wait(i64(tmr.retry_after) * 1000 * time.millisecond)
 		return rest.do(req)
 	}
 	return resp
